Readme/Game/Modularization changes
