+++ package.json
@@ -4,6 +4,7 @@
   "description": "A Node based CLI (Command Line Interface) word guessing game.",
   "main": "index.js",
   "scripts": {
+    "start": "node index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -25,6 +26,7 @@
   },
   "homepage": "https://github.com/Mkkiyoi/WordGuessGame-CLI#readme",
   "dependencies": {
+    "clear": "^0.1.0",
     "inquirer": "^7.0.0"
   }
 }
